requirements.txt
pytorch==1.8.1
numpy
tqdm
toolz 
scikit-learn
scipy
pandas